templates/discord-bot-template/requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
